+++ R/prep_features.R
@@ -4,14 +4,16 @@
 #'geographic, biomes, climate, human footprint.
 #'If desired, bias features need to be calculated separately with ft_bias.
 #'For more control over feature preparation, you can use the
-#'\code{\link{ft_geo}}, \code{\link{ft_biom}}, \code{\link{ft_clim}},\code{\link{ft_foot}} functions.
+#'\code{\link{ft_geo}}, \code{\link{ft_biom}}, \code{\link{ft_clim}},
+#'\code{\link{ft_foot}} functions.
 #'
 #'Without internet access, only geographic features are calculated,
 #'if the sampbias package is not installed, the bias features are skipped.
 #'
 #'@inheritParams ft_geo
 #'@param type character. The type of features to calculate. Possible options are
-#'\dQuote{geographic}, \dQuote{biome}, \dQuote{climate}, \dQuote{human footprint}, \dQuote{bias}.
+#'\dQuote{geographic}, \dQuote{biome}, \dQuote{climate},
+#'\dQuote{human footprint}, \dQuote{bias}.
 #'All except bias are the default.
 #'"\dQuote{bias} is only recommended up to the regional scale or below.
 #'