+++ NEWS.md
@@ -1,4 +1,5 @@
 # IUCNN 0.9.3
+=========================
 
 * add footprint_features function
 * added option for selected variables and ranges to the clim_features function