+++ NEWS.md
@@ -1,3 +1,7 @@
+# IUCNN 0.9.9 (26.05.2021)
+=========================
+* final version for last release tests
+
 # IUCNN 0.9.3
 =========================